Centralize address-write logic and fix admin speedy_automat handling
* Extract shared address sanitization, validation, and payload building into lib/order/address-write

* Fix admin POST/PUT address routes silently mishandling speedy_automat as a street address

* Handle speedy_office and speedy_automat via a single isSpeedyPickup branch across all routes

* Move DB helpers (syncPhoneToProfile, unsetDefaultAddresses) into lib/data/addresses

* Remove cross-route imports of helpers from @/app/api/address/route
